Clip DRAW rows in int32: the last library call in a default-path loop
Reading the renderer objects with nm and objdump, function by function:
every float helper they import comes from the compiled-in float path --
__divsf3 from the float setups and lambdas, sqrtf from the float circle
span, roundf from the float outlines, lrintf from fxFrom -- except three
__divdi3 calls, which were the unrotated DRAW row clip, written in int64
on Thursday and measured flat. A libgcc call per row of every DRAW, on
the default path, invisible to every gate and inside every measurement's
noise. The operands are bounded by the fxFits guard, so int32 gives the
same answers.

  op_draw_asset   9.44 -> 8.90 ms   -6%
  seascape_4     62.18 -> 59.88     -4%
  pixel identity on all three corpora: 21 / 30 / 36; sanitizer clean

The drawing object now imports __divsf3 and sqrtf only, both from the
float path. The default path makes no library call inside any loop.

Recorded alongside: the renderer objects live at
.pio/build/M5Paper_MicroPatterns/src/, outside the env directory and
shared by the normal, bench and profile envs -- correct, since SCons
recompiles on flag change, but it is why alternating envs rebuilds the
renderer, and it is why an earlier search reported no imports at all.

+++ M5Paper_MicroPatterns/src/micropatterns_drawing.cpp
@@ -1513,14 +1513,19 @@ void MicroPatternsDrawing::drawAsset(const DisplayListItem& item, const MicroPat
                     // A rotated asset's box is up to twice its area; every pixel of the
                     // excess used to cost a shift, two compares and a branch to reject.
                     {
-                        const int span = x1 - x0;
-                        const int64_t a0 = axq, d = daxq, top = ((int64_t)aw << MP_FX_SHIFT);
-                        int64_t k0 = 0, k1 = (int64_t)span - 1;      // inclusive run in k
+                        // int32 throughout. The int64 form of this was three __divdi3
+                        // calls -- libgcc, per row of every unrotated DRAW -- and the
+                        // only library call left in the default path's loops. axq is
+                        // bounded by the fxFits guard and aw << 16 by the asset limit,
+                        // so every quantity here fits, and the divide is one instruction.
+                        const int32_t span = x1 - x0;
+                        const int32_t a0 = axq, d = daxq, top = (int32_t)aw << MP_FX_SHIFT;
+                        int32_t k0 = 0, k1 = span - 1;      // inclusive run in k
                         if (d > 0) {
                             if (a0 < 0)     k0 = (-a0 + d - 1) / d;                 // first k with a >= 0
                             k1 = (top - 1 - a0) >= 0 ? (top - 1 - a0) / d : -1;    // last k with a < top
                         } else if (d < 0) {
-                            const int64_t nd = -d;
+                            const int32_t nd = -d;
                             if (a0 >= top)  k0 = (a0 - top + nd) / nd;              // first k with a < top
                             k1 = a0 >= 0 ? a0 / nd : -1;                            // last k with a >= 0
                         } else if (a0 < 0 || a0 >= top) {
@@ -1530,7 +1535,7 @@ void MicroPatternsDrawing::drawAsset(const DisplayListItem& item, const MicroPat
                         if (k1 > span - 1) k1 = span - 1;
                         if (k0 <= k1) {
                             ++_clippedRows;
-                            int32_t a = (int32_t)(a0 + k0 * d);
+                            int32_t a = a0 + k0 * d;
                             const int xs = x0 + (int)k0, xe = x0 + (int)k1 + 1;
                             for (int sx_iter = xs; sx_iter < xe; ++sx_iter) {
                                 if (assetRow[(int)(a >> MP_FX_SHIFT)] == 1) {
